solutions/requirements.txt
azure-ai-projects>=1.0.0b11
azure-identity>=1.20.0
azure-ai-evaluation>=1.5.0
openai>=2.0.0
python-dotenv>=1.0.1
jupyterlab>=4.2.0
ipykernel>=6.29.0
pandas>=2.2.0
